requirements.txt
transformers
PyYAML
pandas
scikit-learn
torch
numpy
scipy
tensorboard==2.2.0
pytorch-lightning==1.2.3
pytorch-nlp==0.5.0
optuna
plotly
sentencepiece
adapter-transformers
captum
entmax